pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptg>=0.5.0.post0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "googletrans==4.0.0rc1",
    "gunicorn>=23.0.0",
    "langdetect>=1.0.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-telegram-bot==13.15",
    "telegram>=0.0.1",
    "telethon>=1.40.0",
]
